refactor: move IDE opening logic to tools/ide.ts
Extract openInIde and its helpers out of the electron shell IPC handler
into a dedicated tools/ide.ts module, leaving shell.ts to only wire the
editor:open-with IPC event. Addresses review feedback.
